+++ tests/unit/agent/test_agent_manager_unit.py
@@ -5,17 +5,19 @@ Unit test AgentManager
 """
 import pytest
 
-from starfish.agent import AgentManager
+from starfish.agent_manager import AgentManager
+from starfish.network.ddo import DDO
 
-def test_agent_manager_add(ethereum_network):
-    manager = AgentManager(ethereum_network)
+def test_agent_manager_add():
+    manager = AgentManager()
 
     authentication = {
         'username': 'user_test',
         'password': 'password_test'
     }
-    manager.add('test', 'http://test.com', authentication=authentication)
-    assert(manager.items['test'])
+    ddo = DDO.create('http://test.com')
+    manager.register_local_agent(ddo, authentication=authentication)
+    assert(manager.local_agent)
 
     with pytest.raises(ValueError):
-        manager.add('failed')
+        manager.register_local_agent('failed')